requirements.txt
Django==4.2.21
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyyaml==6.0.1 